fixed bugs with edit on front-end

+++ components/book/BookBox.js
@@ -10,6 +10,8 @@ var _react2 = _interopRequireDefault(_react);
 
 var _reactRouterDom = require('react-router-dom');
 
+var _auth = require('../../lib/auth');
+
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
 
 function BookBox(_ref) {
@@ -19,15 +21,15 @@ function BookBox(_ref) {
 
   return _react2.default.createElement(
     'article',
-    { className: 'card' },
+    { className: 'card column is-4' },
+    _react2.default.createElement(
+      'h3',
+      { className: 'card-header subtitle' },
+      book.title
+    ),
     _react2.default.createElement(
       _reactRouterDom.Link,
-      { className: 'column is-4 content', to: '/books/' + book._id },
-      _react2.default.createElement(
-        'h3',
-        { className: 'is-subtitle is-3' },
-        book.title
-      ),
+      { className: 'content', to: '/books/' + book._id },
       _react2.default.createElement(
         'p',
         null,
@@ -40,13 +42,21 @@ function BookBox(_ref) {
         book.author
       )
     ),
-    _react2.default.createElement(
-      'a',
-      { className: 'card-footer-item', onClick: function onClick() {
-          return _this.deleteBook(book._id);
-        } },
-      'Delete'
-    )
+    function () {
+      if (book.author === (0, _auth.decodeToken)().username) {
+        return _react2.default.createElement(
+          'footer',
+          { className: 'card-footer-item' },
+          _react2.default.createElement(
+            'button',
+            { className: 'button is-danger', onClick: function onClick() {
+                return _this.deleteBook(book._id);
+              } },
+            'Delete'
+          )
+        );
+      }
+    }()
   );
 }
